removing deleteReferenceToText file and adding functionality to allAnnotations file
